#1、修复数据解析错误

+++ lib/model/rank_project.dart
@@ -1,20 +1,14 @@
-class RankProject{
-  BigInt id;
+class RankProject {
+  int id;
   String name;
   String html_url;
 
-  RankProject({
-    this.id,
-    this.name,
-    this.html_url
-  });
+  RankProject({this.id, this.name, this.html_url});
 
-
-  factory RankProject.fromJson(Map<String, dynamic> parsedJson){
+  factory RankProject.fromJson(Map<String, dynamic> parsedJson) {
     return RankProject(
         id: parsedJson['id'],
-        name : parsedJson['name'],
-        html_url : parsedJson ['html_url']
-    );
+        name: parsedJson['name'],
+        html_url: parsedJson['html_url']);
   }
 }
